feature/98 refactor: config package 위치를 common package 아래로 이동
