+++ src/main/java/com/grasshouse/dorandoran/common/config/jwt/LoginMember.java
@@ -1,4 +1,4 @@
-package com.grasshouse.dorandoran.config.jwt;
+package com.grasshouse.dorandoran.common.config.jwt;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
