+++ src/main/java/com/grasshouse/dorandoran/member/controller/LoginMemberMethodArgumentResolver.java
@@ -2,8 +2,8 @@ package com.grasshouse.dorandoran.member.controller;
 
 import static org.springframework.web.context.request.RequestAttributes.SCOPE_REQUEST;
 
+import com.grasshouse.dorandoran.common.config.jwt.LoginMember;
 import com.grasshouse.dorandoran.common.exception.MemberNotFoundException;
-import com.grasshouse.dorandoran.config.jwt.LoginMember;
 import com.grasshouse.dorandoran.member.repository.MemberRepository;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
