+++ src/main/java/com/grasshouse/dorandoran/post/controller/PostLikeController.java
@@ -1,6 +1,6 @@
 package com.grasshouse.dorandoran.post.controller;
 
-import com.grasshouse.dorandoran.config.jwt.LoginMember;
+import com.grasshouse.dorandoran.common.config.jwt.LoginMember;
 import com.grasshouse.dorandoran.member.domain.Member;
 import com.grasshouse.dorandoran.post.dto.PostLikeCreateRequest;
 import com.grasshouse.dorandoran.post.service.PostLikeService;
