+++ src/test/java/com/grasshouse/dorandoran/member/controller/MemberControllerTest.java
@@ -14,7 +14,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.grasshouse.dorandoran.common.CommonControllerTest;
-import com.grasshouse.dorandoran.config.jwt.JwtTokenProvider;
+import com.grasshouse.dorandoran.common.config.jwt.JwtTokenProvider;
 import com.grasshouse.dorandoran.member.dto.MemberUpdateRequest;
 import com.grasshouse.dorandoran.member.dto.MemberUpdateResponse;
 import com.grasshouse.dorandoran.member.repository.MemberRepository;
